comments addressed

+++ ownership-chain/pallets/asset-metadata-extender/src/types.rs
@@ -11,7 +11,7 @@ pub type UniversalLocationOf<T> = BoundedVec<u8, <T as crate::Config>::MaxUniver
 /// Serves as a position identifier for elements in a collection, facilitating iteration
 /// and element access. It corresponds to the element count, ranging from 0 to N-1 in a collection
 /// of size N.
-pub type Index = u64;
+pub type Index = u32;
 
 /// Asset metadata extension
 /// This will contain the claimer account and location in a raw form
